+++ packages/editor/src/lib/editorStateConversion.js
@@ -1,15 +1,15 @@
 import { convertFromRaw as fromRaw, convertToRaw as toRaw, EditorState } from '@wix/draft-js';
-import { default as Version } from 'wix-rich-content-common/src/Utils/versioningUtils';
+import { version } from '../../package.json';
 
 const addVersion = (obj, version) => {
   obj.VERSION = version;
   return obj;
 };
 
-const convertToRaw = ContentState => addVersion(toRaw(ContentState), Version.getCurrent());
+const convertToRaw = ContentState => addVersion(toRaw(ContentState), version);
 const convertFromRaw = rawState => addVersion(fromRaw(rawState), rawState.VERSION);
 
-const createEmpty = () => addVersion(EditorState.createEmpty(), Version.getCurrent());
+const createEmpty = () => addVersion(EditorState.createEmpty(), version);
 const createWithContent = contentState =>
   addVersion(EditorState.createWithContent(contentState), contentState.VERSION);