+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 > - :house: [Internal]
 
 ## [Unreleased]
+### :bug: Bug Fix
+- `editor`
+  - [#331](https://github.com/wix-incubator/rich-content/pull/331) bad import of current version from common
+  - [#331](https://github.com/wix-incubator/rich-content/pull/331) import only package.json vertios. Instead of the whole package.json.   
 <hr/>
 
 ## 4.0.3 (Jul 10, 2019)